projects/hospital_system: Adding project hospital_system
Files included:
 spp.py
 actions.py

For this project spp.py had to be modified to use a thread for handling
Bluetooth socket events to allow for button input to be processed in
realtime.
Similarly, a timer thread was added to actions.py to use it to alert
the user when a pill is due for administration.