hospital-system: Prevent error with BT invalid codes and alert display
When an invalid code was received, a None response was tried to be
sent, causing an error on server_sock. Now all incorrect codes return
an error message instead. In addition all data is validated as not
None before server_sock.send() is called.

Additionally, the display message for pill administration was not
being shown on the LCD. The order of display was changed to correct
it. Also a minor addition was made to reset the alert timer when the
pill counter is reset.